Add: Add dependency filter functionality
This restructures the original feed-filter into separate files.
Additionally a dependency filter is added. It does not filter based on builtin functions
but based on given scipts and their dependencies.
