+++ rust/src/feed_filter/error.rs
@@ -7,12 +7,14 @@ use std::fmt;
 #[derive(Debug)]
 pub(crate) enum CliError {
     Io(std::io::Error),
+    Message(String),
 }
 
 impl fmt::Display for CliError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
             CliError::Io(error) => write!(f, "{error}"),
+            CliError::Message(msg) => write!(f, "{msg}"),
         }
     }
 }
@@ -20,6 +22,12 @@ impl fmt::Display for CliError {
 impl From<std::io::Error> for CliError {
     fn from(value: std::io::Error) -> Self {
         Self::Io(value)
+    }
+}
+
+impl From<&str> for CliError {
+    fn from(value: &str) -> Self {
+        Self::Message(value.to_string())
     }
 }
